+++ src/targets/operations/open.ts
@@ -46,7 +46,7 @@ export async function execute(context: deploy_targets.TargetOperationExecutionCo
     const WAIT = deploy_helpers.toBooleanSafe(context.operation.wait, true);
 
     await deploy_helpers.open(TARGET_TO_OPEN, {
-        cwd: WORKSPACE.folder.uri.fsPath,
+        cwd: WORKSPACE.rootPath,
         wait: WAIT,
     });
 }